+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Version 0.2.0
 -------------
 
 1. Detect if the download has already been processed. If so skip downloading again and log the skip.
+2. Add the following metrics for Prometheus or other observability tooling
+    1. TODO
+    2. TODO
+    3. TODO
 
 Version 0.1.0
 -------------